Modifications to correct handling of G00 commands
G00 defines moves which tell your machine to go as fast as it can to the next location.  The Laser should be turned off during these moves and the feed rate should be restored following the move.  The original implementation did not handle this correctly.

Corresponding changes were made to the Turnkeylaser inkScape extension.  They are included in the fork since they are also required to work with the changes made to the firmware.